+++ src/global_functions.php
@@ -60,14 +60,14 @@ if (!function_exists('get_child_methods')) {
      * Gets the class methods' names, but unlike the `get_class_methods()`
      *  function, this function excludes the methods of the parent class
      * @param string $class Class name
-     * @return array|null
+     * @return array<string>|null
      * @since 1.0.1
      */
     function get_child_methods(string $class): ?array
     {
         $methods = get_class_methods($class);
-        $parentClass = get_parent_class($class);
 
+        $parentClass = get_parent_class($class);
         if ($parentClass) {
             $methods = array_diff($methods, get_class_methods($parentClass));
         }
@@ -124,7 +124,7 @@ if (!function_exists('is_positive')) {
      */
     function is_positive($string): bool
     {
-        return is_numeric($string) && $string > 0 && $string == round($string);
+        return is_numeric($string) && $string > 0 && $string == round((float)$string);
     }
 }
 
@@ -144,8 +144,8 @@ if (!function_exists('is_stringable')) {
 if (!function_exists('objects_map')) {
     /**
      * Executes an object method for all objects of the given arrays
-     * @param array $objects An array of objects. Each object must have the
-     *  method to be called
+     * @param array<object> $objects An array of objects. Each object must have
+     *  the method to be called
      * @param string $method The method to be called for each object
      * @param array $args Optional arguments for the method to be called
      * @return array Returns an array containing all the returned values of the
@@ -155,7 +155,7 @@ if (!function_exists('objects_map')) {
      */
     function objects_map(array $objects, string $method, array $args = []): array
     {
-        return array_map(function ($object) use ($method, $args) {
+        return array_map(function (object $object) use ($method, $args) {
             Exceptionist::isTrue(method_exists($object, '__call') || method_exists($object, $method), sprintf(
                 'Class `%s` does not have a method `%s`',
                 get_class($object),